requirements.txt
black==23.3.0
bottle==0.12.25
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
clr-loader==0.2.5
colorama==0.4.6
docstring-parser==0.15
greenlet==2.0.2
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
mypy==1.3.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.3
proxy-tools==0.1.0
pycparser==2.21
pythonnet==3.0.1
pywebview==4.1
regex==2023.6.3
requests==2.31.0
SQLAlchemy==2.0.16
typed-argument-parser==1.8.0
typing-inspect==0.9.0
typing_extensions==4.6.3
urllib3==2.0.3
